Move generic keys funcs to util/keys.go
